Factor helper classes out in package.
