+++ neo/rawio/neuralynxrawio/__init__.py
@@ -22,11 +22,10 @@ from neo.rawio.baserawio import (BaseRawIO, _signal_channel_dtype,
 
 import numpy as np
 import os
-import re
-import distutils.version
-import datetime
 from collections import OrderedDict
-import math
+
+from neo.rawio.neuralynxrawio.NcsBlocks import (NcsBlock,NcsBlocksFactory)
+from neo.rawio.neuralynxrawio.NlxHeader import NlxHeader
 
 
 class NeuralynxRawIO(BaseRawIO):
@@ -48,9 +47,8 @@ class NeuralynxRawIO(BaseRawIO):
     extensions = ['nse', 'ncs', 'nev', 'ntt']
     rawmode = 'one-dir'
 
-    _BLOCK_SIZE = 512  # nb sample per signal block
     _ncs_dtype = [('timestamp', 'uint64'), ('channel_id', 'uint32'), ('sample_rate', 'uint32'),
-                  ('nb_valid', 'uint32'), ('samples', 'int16', (_BLOCK_SIZE,))]
+                  ('nb_valid', 'uint32'), ('samples', 'int16', (NcsBlock._BLOCK_SIZE,))]
 
     def __init__(self, dirname='', keep_original_times=False, **kargs):
         """
@@ -349,8 +347,8 @@ class NeuralynxRawIO(BaseRawIO):
         if i_stop is None:
             i_stop = self._sigs_length[seg_index]
 
-        block_start = i_start // self._BLOCK_SIZE
-        block_stop = i_stop // self._BLOCK_SIZE + 1
+        block_start = i_start // NcsBlock._BLOCK_SIZE
+        block_stop = i_stop // NcsBlock._BLOCK_SIZE + 1
         sl0 = i_start % 512
         sl1 = sl0 + (i_stop - i_start)
 
@@ -550,8 +548,8 @@ class NeuralynxRawIO(BaseRawIO):
                     numSampsLastBlock = subdata[-1]['nb_valid']
                     ts0 = subdata[0]['timestamp']
                     ts1 = NcsBlocksFactory.calcSampleTime(curBlocks.sampFreqUsed,
-                                                                      subdata[-1]['timestamp'],
-                                                                      numSampsLastBlock)
+                                                          subdata[-1]['timestamp'],
+                                                          numSampsLastBlock)
                     self._timestamp_limits.append((ts0, ts1))
                     t_start = ts0 / 1e6
                     self._sigs_t_start.append(t_start)
@@ -560,686 +558,8 @@ class NeuralynxRawIO(BaseRawIO):
                     # :TODO: This should really be the total of nb_valid in records, but this
                     #  allows the last record of a block to be shorter, the most common case.
                     #  Have never seen a block of records with not full records before the last.
-                    length = (subdata.size - 1) * self._BLOCK_SIZE + numSampsLastBlock
+                    length = (subdata.size - 1) * NcsBlock._BLOCK_SIZE + numSampsLastBlock
                     self._sigs_length.append(length)
-
-
-class CscRecordHeader:
-    """
-    Information in header of each Ncs record, excluding sample values themselves.
-    """
-
-    def __init__(self, ncsMemMap, recn):
-        """
-        Construct a record header for a given record in a memory map for an NcsFile.
-        """
-        self.timestamp = ncsMemMap['timestamp'][recn]
-        self.channel_id = ncsMemMap['channel_id'][recn]
-        self.sample_rate = ncsMemMap['sample_rate'][recn]
-        self.nb_valid = ncsMemMap['nb_valid'][recn]
-
-
-class NcsBlock:
-    """
-    Information regarding a single contiguous block of records in an Ncs file.
-
-    Model is that times are closed on the left and open on the right. Record
-    numbers are closed on both left and right, that is, inclusive of the last record.
-
-    endTime should never be set less than startTime for comparison functions to work
-    properly, though this is not enforced.
-    """
-
-    def __init__(self):
-        self.startBlock = -1  # index of starting record
-        self.startTime = -1  # starttime of first record
-        self.endBlock = -1  # index of last record (inclusive)
-        self.endTime = -1   # end time of last record, that is, the end time of the last
-                            # sampling period contained in the record
-
-    def __init__(self, sb, st, eb, et):
-        self.startBlock = sb
-        self.startTime = st
-        self.endBlock = eb
-        self.endTime = et
-
-    def before_time(self, rhb):
-        """
-        Determine if this block is completely before another block in time.
-        """
-        return self.endTime < rhb.startTime
-
-    def overlaps_time(self, rhb):
-        """
-        Determine if this block overlaps another in time.
-        """
-        return self.startTime <= rhb.endTime and self.endTime >= rhb.startTime
-
-    def after_time(self, rhb):
-        """
-        Determine if this block is completely after another block in time.
-        """
-        return self.startTime >= rhb.endTime
-
-
-class NcsBlocks:
-    """
-    Contains information regarding the contiguous blocks of records in an Ncs file.
-    Methods of NcsBlocksFactory perform parsing of this information from an Ncs file and
-    produce these where the blocks are discontiguous in time and in temporal order.
-
-    TODO: This class will likely need __eq__, __ne__, and __hash__ to be useful in
-    more sophisticated segment construction algorithms.
-
-    """
-    def __init__(self):
-        self.blocks = []
-        self.sampFreqUsed = 0  # actual sampling frequency of samples
-        self.microsPerSampUsed = 0  # microseconds per sample
-
-
-class NcsBlocksFactory:
-    """
-    Class for factory methods which perform parsing of contiguous blocks of records
-    in Ncs files.
-
-    Model for times is that times are rounded to nearest microsecond. Times
-    from start of a sample until just before the next sample are included,
-    that is, closed lower bound and open upper bound on intervals. A
-    channel with no samples is empty and contains no time intervals.
-
-    Moved here since algorithm covering all 3 header styles and types used is
-    more complicated. Copied from Java code on Sept 7, 2020.
-    """
-
-    _maxGapLength = 5   # maximum gap between predicted and actual block timestamps still
-                        # considered within one NcsBlock
-
-    @staticmethod
-    def getFreqForMicrosPerSamp(micros):
-        """
-        Compute fractional sampling frequency, given microseconds per sample.
-        """
-        return 1e6 / micros
-
-    @staticmethod
-    def getMicrosPerSampForFreq(sampFr):
-        """
-        Calculate fractional microseconds per sample, given the sampling frequency (Hz).
-        """
-        return 1e6 / sampFr
-
-    @staticmethod
-    def calcSampleTime(sampFr, startTime, posn):
-        """
-        Calculate time rounded to microseconds for sample given frequency,
-        start time, and sample position.
-        """
-        return round(startTime + NcsBlocksFactory.getMicrosPerSampForFreq(sampFr) * posn)
-
-    @staticmethod
-    def _parseGivenActualFrequency(ncsMemMap, ncsBlocks, chanNum, reqFreq, blkOnePredTime):
-        """
-        Parse blocks in memory mapped file when microsPerSampUsed and sampFreqUsed are known,
-        filling in an NcsBlocks object.
-
-        PARAMETERS
-        ncsMemMap:
-            memmap of Ncs file
-        ncsBlocks:
-            NcsBlocks with actual sampFreqUsed correct, first NcsBlock with proper startBlock and
-            startTime already added.
-        chanNum:
-            channel number that should be present in all records
-        reqFreq:
-            rounded frequency that all records should contain
-        blkOnePredTime:
-            predicted starting time of second record in block
-
-        RETURN
-        NcsBlocks object with block locations marked
-        """
-        startBlockPredTime = blkOnePredTime
-        blkLen = 0
-        curBlock = ncsBlocks.blocks[0]
-        for recn in range(1, ncsMemMap.shape[0]):
-            hdr = CscRecordHeader(ncsMemMap, recn)
-            if hdr.channel_id != chanNum | hdr.sample_rate != reqFreq:
-                raise IOError('Channel number or sampling frequency changed in ' +
-                              'records within file')
-            predTime = NcsBlocksFactory.calcSampleTime(ncsBlocks.sampFreqUsed,
-                                                                   startBlockPredTime, blkLen)
-            nValidSamps = hdr.nb_valid
-            if hdr.timestamp != predTime:
-                curBlock.endBlock = recn - 1
-                curBlock.endTime = predTime
-                curBlock = NcsBlock(recn, hdr.timestamp, -1, -1)
-                ncsBlocks.blocks.append(curBlock)
-                startBlockPredTime = NcsBlocksFactory.calcSampleTime(
-                    ncsBlocks.sampFreqUsed,
-                    hdr.timestamp,
-                    nValidSamps)
-                blkLen = 0
-            else:
-                blkLen += nValidSamps
-
-        curBlock.endBlock = ncsMemMap.shape[0] - 1
-        endTime = NcsBlocksFactory.calcSampleTime(ncsBlocks.sampFreqUsed,
-                                                              startBlockPredTime,
-                                                              blkLen)
-        curBlock.endTime = endTime
-
-        return ncsBlocks
-
-    @staticmethod
-    def _buildGivenActualFrequency(ncsMemMap, actualSampFreq, reqFreq):
-        """
-        Build NcsBlocks object for file given actual sampling frequency.
-
-        Requires that frequency in each record agrees with requested frequency. This is
-        normally obtained by rounding the header frequency; however, this value may be different
-        from the rounded actual frequency used in the recording, since the underlying
-        requirement in older Ncs files was that the number of microseconds per sample in the
-        records is the inverse of the sampling frequency stated in the header truncated to
-        whole microseconds.
-
-        PARAMETERS
-        ncsMemMap:
-            memmap of Ncs file
-        actualSampFreq:
-            actual sampling frequency used
-        reqFreq:
-            frequency to require in records
-
-        RETURN:
-            NcsBlocks object
-        """
-        # check frequency in first record
-        rh0 = CscRecordHeader(ncsMemMap, 0)
-        if rh0.sample_rate != reqFreq:
-            raise IOError("Sampling frequency in first record doesn't agree with header.")
-        chanNum = rh0.channel_id
-
-        nb = NcsBlocks()
-        nb.sampFreqUsed = actualSampFreq
-        nb.microsPerSampUsed = NcsBlocksFactory.getMicrosPerSampForFreq(actualSampFreq)
-
-        # check if file is one block of records, which is often the case, and avoid full parse
-        lastBlkI = ncsMemMap.shape[0] - 1
-        rhl = CscRecordHeader(ncsMemMap, lastBlkI)
-        predLastBlockStartTime = NcsBlocksFactory.calcSampleTime(actualSampFreq,
-                                                                             rh0.timestamp,
-                                                        NeuralynxRawIO._BLOCK_SIZE * lastBlkI)
-        if rhl.channel_id == chanNum and rhl.sample_rate == reqFreq and \
-                rhl.timestamp == predLastBlockStartTime:
-            lastBlkEndTime = NcsBlocksFactory.calcSampleTime(actualSampFreq,
-                                                                         rhl.timestamp,
-                                                                         rhl.nb_valid)
-            curBlock = NcsBlock(0, rh0.timestamp, lastBlkI, lastBlkEndTime)
-
-            nb.blocks.append(curBlock)
-            return nb
-
-        # otherwise need to scan looking for breaks
-        else:
-            blkOnePredTime = NcsBlocksFactory.calcSampleTime(actualSampFreq,
-                                                                         rh0.timestamp,
-                                                                         rh0.nb_valid)
-            curBlock = NcsBlock(0, rh0.timestamp, -1, -1)
-            nb.blocks.append(curBlock)
-            return NcsBlocksFactory._parseGivenActualFrequency(ncsMemMap, nb, chanNum, reqFreq,
-                                                               blkOnePredTime)
-
-    @staticmethod
-    def _parseForMaxGap(ncsMemMap, ncsBlocks, maxGapLen):
-        """
-        Parse blocks of records from file, allowing a maximum gap in timestamps between records
-        in blocks. Estimates frequency being used based on timestamps.
-
-        PARAMETERS
-        ncsMemMap:
-            memmap of Ncs file
-        ncsBlocks:
-            NcsBlocks object with sampFreqUsed set to nominal frequency to use in computing time
-            for samples (Hz)
-        maxGapLen:
-            maximum difference within a block between predicted time of start of record and
-            recorded time
-
-        RETURN:
-            NcsBlocks object with sampFreqUsed and microsPerSamp set based on estimate from
-            largest block
-        """
-
-        # track frequency of each block and use estimate with longest block
-        maxBlkLen = 0
-        maxBlkFreqEstimate = 0
-
-        # Parse the record sequence, finding blocks of continuous time with no more than
-        # maxGapLength and same channel number
-        rh0 = CscRecordHeader(ncsMemMap, 0)
-        chanNum = rh0.channel_id
-
-        startBlockTime = rh0.timestamp
-        blkLen = rh0.nb_valid
-        lastRecTime = rh0.timestamp
-        lastRecNumSamps = rh0.nb_valid
-        recFreq = rh0.sample_rate
-
-        curBlock = NcsBlock(0, rh0.timestamp, -1, -1)
-        ncsBlocks.blocks.append(curBlock)
-        for recn in range(1, ncsMemMap.shape[0]):
-            hdr = CscRecordHeader(ncsMemMap, recn)
-            if hdr.channel_id != chanNum or hdr.sample_rate != recFreq:
-                raise IOError('Channel number or sampling frequency changed in ' +
-                              'records within file')
-            predTime = NcsBlocksFactory.calcSampleTime(ncsBlocks.sampFreqUsed,
-                                                                   lastRecTime, lastRecNumSamps)
-            if abs(hdr.timestamp - predTime) > maxGapLen:
-                curBlock.endBlock = recn - 1
-                curBlock.endTime = predTime
-                curBlock = NcsBlock(recn, hdr.timestamp, -1, -1)
-                ncsBlocks.blocks.append(curBlock)
-                if blkLen > maxBlkLen:
-                    maxBlkLen = blkLen
-                    maxBlkFreqEstimate = (blkLen - lastRecNumSamps) * 1e6 / \
-                                         (lastRecTime - startBlockTime)
-                startBlockTime = hdr.timestamp
-                blkLen = hdr.nb_valid
-            else:
-                blkLen += hdr.nb_valid
-            lastRecTime = hdr.timestamp
-            lastRecNumSamps = hdr.nb_valid
-
-        curBlock.endBlock = ncsMemMap.shape[0] - 1
-        endTime = NcsBlocksFactory.calcSampleTime(ncsBlocks.sampFreqUsed,
-                                                              lastRecTime, lastRecNumSamps)
-        curBlock.endTime = endTime
-
-        ncsBlocks.sampFreqUsed = maxBlkFreqEstimate
-        ncsBlocks.microsPerSampUsed = NcsBlocksFactory.getMicrosPerSampForFreq(maxBlkFreqEstimate)
-
-        return ncsBlocks
-
-    @staticmethod
-    def _buildForMaxGap(ncsMemMap, nomFreq):
-        """
-        Determine blocks of records in memory mapped Ncs file given a nominal frequency of
-        the file, using the default values of frequency tolerance and maximum gap between blocks.
-
-        PARAMETERS
-        ncsMemMap:
-            memmap of Ncs file
-        nomFreq:
-            nominal sampling frequency used, normally from header of file
-
-        RETURN:
-            NcsBlocks object
-        """
-        nb = NcsBlocks()
-
-        numRecs = ncsMemMap.shape[0]
-        if numRecs < 1:
-            return nb
-
-        rh0 = CscRecordHeader(ncsMemMap, 0)
-        chanNum = rh0.channel_id
-
-        lastBlkI = numRecs - 1
-        rhl = CscRecordHeader(ncsMemMap, lastBlkI)
-
-        # check if file is one block of records, with exact timestamp match, which may be the case
-        numSampsForPred = NeuralynxRawIO._BLOCK_SIZE * lastBlkI
-        predLastBlockStartTime = NcsBlocksFactory.calcSampleTime(nomFreq,
-                                                                             rh0.timestamp,
-                                                                             numSampsForPred)
-        freqInFile = math.floor(nomFreq)
-        if abs(rhl.timestamp - predLastBlockStartTime) == 0 and \
-                rhl.channel_id == chanNum and rhl.sample_rate == freqInFile:
-            endTime = NcsBlocksFactory.calcSampleTime(nomFreq, rhl.timestamp, rhl.nb_valid)
-            curBlock = NcsBlock(0, rh0.timestamp, lastBlkI, endTime)
-            nb.blocks.append(curBlock)
-            nb.sampFreqUsed = numSampsForPred / (rhl.timestamp - rh0.timestamp) * 1e6
-            nb.microsPerSampUsed = NcsBlocksFactory.getMicrosPerSampForFreq(nb.sampFreqUsed)
-
-        # otherwise parse records to determine blocks using default maximum gap length
-        else:
-            nb.sampFreqUsed = nomFreq
-            nb.microsPerSampUsed = NcsBlocksFactory.getMicrosPerSampForFreq(nb.sampFreqUsed)
-            nb = NcsBlocksFactory._parseForMaxGap(ncsMemMap, nb, NcsBlocksFactory._maxGapLength)
-
-        return nb
-
-    @staticmethod
-    def buildForNcsFile(ncsMemMap, nlxHdr):
-        """
-        Build an NcsBlocks object for an NcsFile, given as a memmap and NlxHeader,
-        handling gap detection appropriately given the file type as specified by the header.
-
-        PARAMETERS
-        ncsMemMap:
-            memory map of file
-        acqType:
-            string specifying type of data acquisition used, one of types returned by
-            NlxHeader.typeOfRecording()
-        """
-        acqType = nlxHdr.type_of_recording()
-
-        # Old Neuralynx style with truncated whole microseconds for actual sampling. This
-        # restriction arose from the sampling being based on a master 1 MHz clock.
-        if acqType == "PRE4":
-            freq = nlxHdr['sampling_rate']
-            microsPerSampUsed = math.floor(NcsBlocksFactory.getMicrosPerSampForFreq(freq))
-            sampFreqUsed = NcsBlocksFactory.getFreqForMicrosPerSamp(microsPerSampUsed)
-            nb = NcsBlocksFactory._buildGivenActualFrequency(ncsMemMap, sampFreqUsed, math.floor(freq))
-            nb.sampFreqUsed = sampFreqUsed
-            nb.microsPerSampUsed = microsPerSampUsed
-
-        # digital lynx style with fractional frequency and micros per samp determined from
-        # block times
-        elif acqType == "DIGITALLYNX" or acqType == "DIGITALLYNXSX":
-            nomFreq = nlxHdr['sampling_rate']
-            nb = NcsBlocksFactory._buildForMaxGap(ncsMemMap, nomFreq)
-
-        # BML style with fractional frequency and micros per samp
-        elif acqType == "BML":
-            sampFreqUsed = nlxHdr['sampling_rate']
-            nb = NcsBlocksFactory._buildGivenActualFrequency(ncsMemMap, sampFreqUsed,
-                                                             math.floor(sampFreqUsed))
-
-        else:
-            raise TypeError("Unknown Ncs file type from header.")
-
-        return nb
-
-    @staticmethod
-    def _verifyBlockStructure(ncsMemMap, ncsBlocks):
-        """
-        Check that the record structure and timestamps for the ncsMemMap and nlxHeader
-        agrees with that in ncsBlocks.
-
-        Provides a more rapid verification of struture than building a new NcsBlocks
-        and checking equality.
-
-        PARAMETERS
-        ncsMemMap:
-            memmap of file to be checked
-        ncsBlocks
-            existing block structure to be checked
-        RETURN:
-            true if all timestamps and block record starts and stops agree, otherwise false.
-        """
-        for blki in range(0, len(ncsBlocks.blocks)):
-            stHdr = CscRecordHeader(ncsMemMap, ncsBlocks.blocks[blki].startBlock)
-            if stHdr.timestamp != ncsBlocks.blocks[blki].startTime: return False
-            endHdr = CscRecordHeader(ncsMemMap, ncsBlocks.blocks[blki].endBlock)
-            endTime = NcsBlocksFactory.calcSampleTime(ncsBlocks.sampFreqUsed,
-                                                                  endHdr.timestamp,
-                                                                  endHdr.nb_valid)
-            if endTime != ncsBlocks.blocks[blki].endTime: return False
-
-        return True
-
-
-class NlxHeader(OrderedDict):
-    """
-    Representation of basic information in all 16 kbytes Neuralynx file headers,
-    including dates opened and closed if given.
-    """
-
-    HEADER_SIZE = 2 ** 14  # Neuralynx files have a txt header of 16kB
-
-    # helper function to interpret boolean keys
-    def _to_bool(txt):
-        if txt == 'True':
-            return True
-        elif txt == 'False':
-            return False
-        else:
-            raise Exception('Can not convert %s to bool' % txt)
-
-    # keys that may be present in header which we parse
-    txt_header_keys = [
-        ('AcqEntName', 'channel_names', None),  # used
-        ('FileType', '', None),
-        ('FileVersion', '', None),
-        ('RecordSize', '', None),
-        ('HardwareSubSystemName', '', None),
-        ('HardwareSubSystemType', '', None),
-        ('SamplingFrequency', 'sampling_rate', float),  # used
-        ('ADMaxValue', '', None),
-        ('ADBitVolts', 'bit_to_microVolt', None),  # used
-        ('NumADChannels', '', None),
-        ('ADChannel', 'channel_ids', None),  # used
-        ('InputRange', '', None),
-        ('InputInverted', 'input_inverted', _to_bool),  # used
-        ('DSPLowCutFilterEnabled', '', None),
-        ('DspLowCutFrequency', '', None),
-        ('DspLowCutNumTaps', '', None),
-        ('DspLowCutFilterType', '', None),
-        ('DSPHighCutFilterEnabled', '', None),
-        ('DspHighCutFrequency', '', None),
-        ('DspHighCutNumTaps', '', None),
-        ('DspHighCutFilterType', '', None),
-        ('DspDelayCompensation', '', None),
-        ('DspFilterDelay_µs', '', None),
-        ('DisabledSubChannels', '', None),
-        ('WaveformLength', '', int),
-        ('AlignmentPt', '', None),
-        ('ThreshVal', '', None),
-        ('MinRetriggerSamples', '', None),
-        ('SpikeRetriggerTime', '', None),
-        ('DualThresholding', '', None),
-        (r'Feature \w+ \d+', '', None),
-        ('SessionUUID', '', None),
-        ('FileUUID', '', None),
-        ('CheetahRev', '', None),  # only for older versions of Cheetah
-        ('ProbeName', '', None),
-        ('OriginalFileName', '', None),
-        ('TimeCreated', '', None),
-        ('TimeClosed', '', None),
-        ('ApplicationName', '', None),  # also include version number when present
-        ('AcquisitionSystem', '', None),
-        ('ReferenceChannel', '', None),
-        ('NLX_Base_Class_Type', '', None)  # in version 4 and earlier versions of Cheetah
-    ]
-
-    # Filename and datetime may appear in header lines starting with # at
-    # beginning of header or in later versions as a property. The exact format
-    # used depends on the application name and its version as well as the
-    # -FileVersion property.
-    #
-    # There are 4 styles understood by this code and the patterns used for parsing
-    # the items within each are stored in a dictionary. Each dictionary is then
-    # stored in main dictionary keyed by an abbreviation for the style.
-    header_pattern_dicts = {
-        # BML
-        'bml': dict(
-            datetime1_regex=r'## Time Opened: \(m/d/y\): (?P<date>\S+)'
-                            r'  At Time: (?P<time>\S+)',
-            filename_regex=r'## File Name: (?P<filename>\S+)',
-            datetimeformat='%m/%d/%y %H:%M:%S.%f'),
-        # Cheetah after version 1 and before version 5
-        'bv5': dict(
-            datetime1_regex=r'## Time Opened: \(m/d/y\): (?P<date>\S+)'
-                            r'  At Time: (?P<time>\S+)',
-            filename_regex=r'## File Name: (?P<filename>\S+)',
-            datetimeformat='%m/%d/%Y %H:%M:%S.%f'),
-        # Cheetah version 5 before and including v 5.6.4 as well as version 1
-        'bv5.6.4': dict(
-            datetime1_regex=r'## Time Opened \(m/d/y\): (?P<date>\S+)'
-                            r'  \(h:m:s\.ms\) (?P<time>\S+)',
-            datetime2_regex=r'## Time Closed \(m/d/y\): (?P<date>\S+)'
-                            r'  \(h:m:s\.ms\) (?P<time>\S+)',
-            filename_regex=r'## File Name (?P<filename>\S+)',
-            datetimeformat='%m/%d/%Y %H:%M:%S.%f'),
-        # Cheetah after v 5.6.4 and default for others such as Pegasus
-        'def': dict(
-            datetime1_regex=r'-TimeCreated (?P<date>\S+) (?P<time>\S+)',
-            datetime2_regex=r'-TimeClosed (?P<date>\S+) (?P<time>\S+)',
-            filename_regex=r'-OriginalFileName "?(?P<filename>\S+)"?',
-            datetimeformat='%Y/%m/%d %H:%M:%S')
-    }
-
-    def build_for_file(filename):
-        """
-        Factory function to build NlxHeader for a given file.
-        """
-
-        with open(filename, 'rb') as f:
-            txt_header = f.read(NlxHeader.HEADER_SIZE)
-        txt_header = txt_header.strip(b'\x00').decode('latin-1')
-
-        # must start with 8 # characters
-        assert txt_header.startswith("########"),\
-            'Neuralynx files must start with 8 # characters.'
-
-        # find keys
-        info = NlxHeader()
-        for k1, k2, type_ in NlxHeader.txt_header_keys:
-            pattern = r'-(?P<name>' + k1 + r')\s+(?P<value>[\S ]*)'
-            matches = re.findall(pattern, txt_header)
-            for match in matches:
-                if k2 == '':
-                    name = match[0]
-                else:
-                    name = k2
-                value = match[1].rstrip(' ')
-                if type_ is not None:
-                    value = type_(value)
-                info[name] = value
-
-        # if channel_ids or s not in info then the filename is used
-        name = os.path.splitext(os.path.basename(filename))[0]
-
-        # convert channel ids
-        if 'channel_ids' in info:
-            chid_entries = re.findall(r'\w+', info['channel_ids'])
-            info['channel_ids'] = [int(c) for c in chid_entries]
-        else:
-            info['channel_ids'] = [name]
-
-        # convert channel names
-        if 'channel_names' in info:
-            name_entries = re.findall(r'\w+', info['channel_names'])
-            if len(name_entries) == 1:
-                info['channel_names'] = name_entries * len(info['channel_ids'])
-            assert len(info['channel_names']) == len(info['channel_ids']), \
-                'Number of channel ids does not match channel names.'
-        else:
-            info['channel_names'] = [name] * len(info['channel_ids'])
-
-        # version and application name
-        # older Cheetah versions with CheetahRev property
-        if 'CheetahRev' in info:
-            assert 'ApplicationName' not in info
-            info['ApplicationName'] = 'Cheetah'
-            app_version = info['CheetahRev']
-        # new file version 3.4 does not contain CheetahRev property, but ApplicationName instead
-        elif 'ApplicationName' in info:
-            pattern = r'(\S*) "([\S ]*)"'
-            match = re.findall(pattern, info['ApplicationName'])
-            assert len(match) == 1, 'impossible to find application name and version'
-            info['ApplicationName'], app_version = match[0]
-        # BML Ncs file writing contained neither property, assume BML version 2
-        else:
-            info['ApplicationName'] = 'BML'
-            app_version = "2.0"
-
-        info['ApplicationVersion'] = distutils.version.LooseVersion(app_version)
-
-        # convert bit_to_microvolt
-        if 'bit_to_microVolt' in info:
-            btm_entries = re.findall(r'\S+', info['bit_to_microVolt'])
-            if len(btm_entries) == 1:
-                btm_entries = btm_entries * len(info['channel_ids'])
-            info['bit_to_microVolt'] = [float(e) * 1e6 for e in btm_entries]
-            assert len(info['bit_to_microVolt']) == len(info['channel_ids']), \
-                'Number of channel ids does not match bit_to_microVolt conversion factors.'
-
-        if 'InputRange' in info:
-            ir_entries = re.findall(r'\w+', info['InputRange'])
-            if len(ir_entries) == 1:
-                info['InputRange'] = [int(ir_entries[0])] * len(chid_entries)
-            else:
-                info['InputRange'] = [int(e) for e in ir_entries]
-            assert len(info['InputRange']) == len(chid_entries), \
-                'Number of channel ids does not match input range values.'
-
-        # Format of datetime depends on app name, app version
-        # :TODO: this works for current examples but is not likely actually related
-        # to app version in this manner.
-        an = info['ApplicationName']
-        if an == 'Cheetah':
-            av = info['ApplicationVersion']
-            mv = av.version[0]
-            if av <= '2':  # version 1 uses same as older versions
-                hpd = NlxHeader.header_pattern_dicts['bv5.6.4']
-            elif av < '5':
-                hpd = NlxHeader.header_pattern_dicts['bv5']
-            elif av <= '5.6.4':
-                hpd = NlxHeader.header_pattern_dicts['bv5.6.4']
-            else:
-                hpd = NlxHeader.header_pattern_dicts['def']
-        elif an == 'BML':
-            hpd = NlxHeader.header_pattern_dicts['bml']
-        else:
-            hpd = NlxHeader.header_pattern_dicts['def']
-
-        # opening time
-        dt1 = re.search(hpd['datetime1_regex'], txt_header).groupdict()
-        info['recording_opened'] = datetime.datetime.strptime(
-            dt1['date'] + ' ' + dt1['time'], hpd['datetimeformat'])
-
-        # close time, if available
-        if 'datetime2_regex' in hpd:
-            dt2 = re.search(hpd['datetime2_regex'], txt_header).groupdict()
-            info['recording_closed'] = datetime.datetime.strptime(
-                dt2['date'] + ' ' + dt2['time'], hpd['datetimeformat'])
-
-        return info
-
-    def type_of_recording(self):
-        """
-        Determines type of recording in Ncs file with this header.
-
-        RETURN:
-            one of 'PRE4','BML','DIGITALLYNX','DIGITALLYNXSX','UNKNOWN'
-        """
-
-        if 'NLX_Base_Class_Type' in self:
-
-            # older style standard neuralynx acquisition with rounded sampling frequency
-            if self['NLX_Base_Class_Type'] == 'CscAcqEnt':
-                return 'PRE4'
-
-            # BML style with fractional frequency and microsPerSamp
-            elif self['NLX_Base_Class_Type'] == 'BmlAcq':
-                return 'BML'
-
-            else:
-                return 'UNKNOWN'
-
-        elif 'HardwareSubSystemType' in self:
-
-            # DigitalLynx
-            if self['HardwareSubSystemType'] == 'DigitalLynx':
-                return 'DIGITALLYNX'
-
-            # DigitalLynxSX
-            elif self['HardwareSubSystemType'] == 'DigitalLynxSX':
-                return 'DIGITALLYNXSX'
-
-        elif 'FileType' in self:
-
-            if self['FileVersion'] in ['3.3', '3.4']:
-                return self['AcquisitionSystem'].split()[1].upper()
-
-            else:
-                return 'UNKNOWN'
-
-        else:
-            return 'UNKNOWN'
 
 
 nev_dtype = [
